refactor(ui): move header folder to ui folder
